requirements.txt
pandas==2.2.3
google-api-python-client

python-dotenv==1.0.0
requests
langchain_groq
selenium
beautifulsoup4